test: add intg test for nip-33
